Refactored website-specific implementation code into Website module. Included index files
